[example] More APIs example for audio effect/mixing
