+++ example/lib/examples/advanced/index.dart
@@ -17,7 +17,7 @@ import 'package:agora_rtc_engine_example/examples/advanced/stream_message/stream
 import 'package:agora_rtc_engine_example/examples/advanced/take_snapshot/take_snapshot.dart';
 import 'package:flutter/foundation.dart';
 
-import 'audio_mixing/audio_mixing.dart';
+import 'audio_effect_mixing/audio_effect_mixing.dart';
 import 'audio_spectrum/audio_spectrum.dart';
 import 'channel_media_relay/channel_media_relay.dart';
 import 'device_manager/device_manager.dart';
@@ -36,7 +36,7 @@ import 'voice_changer/voice_changer.dart';
 /// Data source for advanced examples
 final advanced = [
   {'name': 'Advanced'},
-  if (!kIsWeb) {'name': 'AudioMixing', 'widget': const AudioMixing()},
+  if (!kIsWeb) {'name': 'AudioEffectMixing', 'widget': const AudioEffectMixing()},
   if (!kIsWeb)
     {'name': 'ChannelMediaRelay', 'widget': const ChannelMediaRelay()},
   if (kIsWeb || !(Platform.isAndroid || Platform.isIOS))
