+++ example/lib/examples/basic/join_channel_audio/join_channel_audio.dart
@@ -160,30 +160,6 @@ class _State extends State<JoinChannelAudio> {
     });
   }
 
-  _switchEffect() async {
-    if (playEffect) {
-      await _engine.stopEffect(1);
-      setState(() {
-        playEffect = false;
-      });
-    } else {
-      final path =
-          (await _engine.getAssetAbsolutePath("assets/Sound_Horizon.mp3"))!;
-      await _engine.playEffect(
-          soundId: 1,
-          filePath: path,
-          loopCount: 0,
-          pitch: 1,
-          pan: 1,
-          gain: 100,
-          publish: true);
-      // .then((value) {
-      setState(() {
-        playEffect = true;
-      });
-    }
-  }
-
   _onChangeInEarMonitoringVolume(double value) async {
     _inEarMonitoringVolume = value;
     await _engine.setInEarMonitoringVolume(_inEarMonitoringVolume.toInt());
@@ -261,7 +237,7 @@ class _State extends State<JoinChannelAudio> {
                       uid: int.tryParse(_selectedUidController.text) ?? -1,
                       mute: true);
                 },
-                child: Text('mute Remote Audio'),
+                child: const Text('mute Remote Audio'),
               ),
               ElevatedButton(
                 onPressed: () async {
@@ -269,7 +245,7 @@ class _State extends State<JoinChannelAudio> {
                       uid: int.tryParse(_selectedUidController.text) ?? -1,
                       mute: false);
                 },
-                child: Text('unmute Remote Audio'),
+                child: const Text('unmute Remote Audio'),
               ),
             ],
           ],
@@ -315,11 +291,6 @@ class _State extends State<JoinChannelAudio> {
                       child: Text(
                           enableSpeakerphone ? 'Speakerphone' : 'Earpiece'),
                     ),
-                    if (!kIsWeb)
-                      ElevatedButton(
-                        onPressed: isJoined ? _switchEffect : null,
-                        child: Text('${playEffect ? 'Stop' : 'Play'} effect'),
-                      ),
                     Row(
                       mainAxisAlignment: MainAxisAlignment.end,
                       children: [
